pyproject.toml
[tool.poetry]
name = "proxmox-cluster-tasks"
version = "0.1.3"
description = ""
authors = ["lexxai <lexxai@gmail.com>"]
readme = "README.md"
packages = [{include = "cluster_tasks", from = "src"}]


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "==1.0.1"
httpx = {extras = ["http2"], version = "^0.28.1"}
paramiko = "^3.5.0"
asyncssh = "^2.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
proxmox-cluster-tasks-cli = "cluster_tasks.cli:main"
proxmox-cluster-tasks-api = "cluster_tasks.api:main"

# Correct placement for package inclusion
